@@ -2,8 +2,8 @@ package nilherman.funfacts.presentation
 
 import android.os.Build
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
 import nilherman.funfacts.R
 import nilherman.funfacts.presentation.detail.DetailsFragment
 
