@@ -3,7 +3,7 @@ package nilherman.funfacts.presentation
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import nilherman.funfacts.R
 import nilherman.funfacts.presentation.home.MainActivity
 
