@@ -2,24 +2,25 @@ package nilherman.funfacts.presentation.home.characters
 
 import android.content.Context
 import android.os.Bundle
-import android.support.v7.widget.LinearLayoutManager
-import android.util.Log
 import android.view.*
-import kotlinx.android.synthetic.main.fragment_list.*
 import nilherman.funfacts.R
 import nilherman.funfacts.SUPERHERO
-import nilherman.funfacts.data.apiclient.Repository
-import nilherman.funfacts.data.apiclient.RestApi
-import nilherman.funfacts.data.model.characters.Response
 import nilherman.funfacts.data.model.characters.ResultsItem
 import nilherman.funfacts.presentation.BaseFragment
-import retrofit2.Call
-import retrofit2.Callback
-import android.support.v7.widget.SearchView
 import android.view.inputmethod.InputMethodManager
+import androidx.appcompat.widget.SearchView
+import androidx.lifecycle.ViewModelProviders
+import androidx.recyclerview.widget.LinearLayoutManager
+import kotlinx.android.synthetic.main.fragment_list.*
 import kotlinx.android.synthetic.main.fragment_list.view.*
+import nilherman.funfacts.data.apiclient.MarvelApiImpl
+import nilherman.funfacts.data.apiclient.Repository
+import nilherman.funfacts.data.apiclient.RestApi
+import nilherman.funfacts.domain.characters.CharactersViewModel
 
-class CharactersFragment : BaseFragment() , Callback<Response> {
+class CharactersFragment : BaseFragment() {
+
+    private lateinit var viewModel: CharactersViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -29,7 +30,6 @@ class CharactersFragment : BaseFragment() , Callback<Response> {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         val view = inflater.inflate(R.layout.fragment_list, container, false)
         setHasOptionsMenu(true)
-
         initRecyclerView(view)
 
         return view
@@ -37,25 +37,18 @@ class CharactersFragment : BaseFragment() , Callback<Response> {
 
     override fun onStart() {
         super.onStart()
-        //initListeners()
-        Repository().getCharacter(RestApi.apikey, RestApi.ts, RestApi.hash).enqueue(this)
+        activity?.let {activity ->
+            viewModel = ViewModelProviders.of(this).get(CharactersViewModel::class.java)
+            viewModel.getCharacters().observe(this, androidx.lifecycle.Observer { characters ->
+                recyclerView.adapter = CharactersAdapter(characters = characters.data?.results, context = activity) {
+                    character: ResultsItem -> onCharactersClicked(character)
+                }
+            })
+        }
     }
 
     private fun initRecyclerView(view : View) {
         view.recyclerView.layoutManager = LinearLayoutManager(requireContext())
-    }
-
-    override fun onFailure(call: Call<Response>?, t: Throwable?) {
-        Log.d("", t?.message.toString())
-    }
-
-    override fun onResponse(call: Call<Response>?, response: retrofit2.Response<Response>?) {
-        context?.let { context ->
-            recyclerView.adapter = CharactersAdapter(response?.body()?.data?.results, context) {
-                character: ResultsItem -> onCharactersClicked(character)
-            }
-            recyclerView.adapter?.notifyDataSetChanged()
-        }
     }
 
     private fun onCharactersClicked(character : ResultsItem) {
@@ -76,24 +69,27 @@ class CharactersFragment : BaseFragment() , Callback<Response> {
 
         val searchItem = menu?.findItem(R.id.action_search)
         val searchView = searchItem?.actionView as SearchView
+        searchView.queryHint = getString(R.string.search_hint)
 
         fun doSearch() {
             val searchFilter = searchView.toString()
+            val marvelApi = MarvelApiImpl()
             if (searchFilter.isEmpty()) {
-                Repository().getCharacter(RestApi.apikey, RestApi.ts, RestApi.hash).enqueue(this)
+                marvelApi.getCharacter(RestApi.apikey, RestApi.ts, RestApi.hash)
             } else {
-                Repository().getCharacter(RestApi.apikey, RestApi.ts, RestApi.hash, searchFilter).enqueue(this)
+                marvelApi.getCharacter(RestApi.apikey, RestApi.ts, RestApi.hash, searchFilter)
             }
         }
 
         searchView.setOnQueryTextListener(object :  SearchView.OnQueryTextListener {
             override fun onQueryTextSubmit(query: String?): Boolean {
+                doSearch()
                 searchView.closeKeyboard()
                 return false
             }
 
             override fun onQueryTextChange(newText: String?): Boolean {
-                doSearch()
+
                 return false
             }
         })
